clean up DB code by creating models.py
